Remove AppController #287
- move methods to Web\Controller
